[Ojas] feat: Implement Full CRUD Operations for Employee Payroll with MySQL

+++ EmployeePayrollApp/src/main/java/com/bridgelabz/EmployeePayrollApp/repository/EmployeePayrollRepository.java
@@ -3,5 +3,5 @@ package com.bridgelabz.EmployeePayrollApp.repository;
 import com.bridgelabz.EmployeePayrollApp.model.Employee;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EmployeePayrollRepository extends JpaRepository<Employee, Integer> {
+public interface EmployeePayrollRepository extends JpaRepository<Employee, Long> {
 }
